fix arm64;

+++ src/runtime/Classes/PBArray.h
@@ -41,30 +41,30 @@ typedef enum _PBArrayValueType
 {
 @protected
 	PBArrayValueType	_valueType;
-	unsigned long			_capacity;
-	unsigned long			_count;
+	NSUInteger			_capacity;
+	NSUInteger			_count;
 	void *				_data;
 
 }
 
-- (unsigned long)count;
-- (id)objectAtIndex:(unsigned long)index;
-- (BOOL)boolAtIndex:(unsigned long)index;
-- (long)int32AtIndex:(unsigned long)index;
-- (unsigned long)uint32AtIndex:(unsigned long)index;
-- (long long)int64AtIndex:(unsigned long)index;
-- (unsigned long long)uint64AtIndex:(unsigned long)index;
-- (Float32)floatAtIndex:(unsigned long)index;
-- (Float64)doubleAtIndex:(unsigned long)index;
+- (NSUInteger)count;
+- (id)objectAtIndex:(NSUInteger)index;
+- (BOOL)boolAtIndex:(NSUInteger)index;
+- (long)int32AtIndex:(NSUInteger)index;
+- (NSUInteger)uint32AtIndex:(NSUInteger)index;
+- (long long)int64AtIndex:(NSUInteger)index;
+- (unsigned long long)uint64AtIndex:(NSUInteger)index;
+- (Float32)floatAtIndex:(NSUInteger)index;
+- (Float64)doubleAtIndex:(NSUInteger)index;
 - (BOOL)isEqualToArray:(PBArray *)array;
-- (void)enumerateObjectsUsingBlock:(void (^)(id obj, unsigned long idx, BOOL *stop))block;
+- (void)enumerateObjectsUsingBlock:(void (^)(id obj, NSUInteger idx, BOOL *stop))block;
 - (id)firstObject;
 - (id)lastObject;
-- (id)objectAtIndexedSubscript:(unsigned long)idx;
+- (id)objectAtIndexedSubscript:(NSUInteger)idx;
 
 @property (nonatomic,assign,readonly) PBArrayValueType valueType;
 @property (nonatomic,assign,readonly) const void * data;
-@property (nonatomic,assign,readonly,getter=count) unsigned long count;
+@property (nonatomic,assign,readonly,getter=count) NSUInteger count;
 
 @end
 
@@ -77,10 +77,10 @@ typedef enum _PBArrayValueType
 @interface PBArray (PBArrayCreation)
 
 + (id)arrayWithValueType:(PBArrayValueType)valueType;
-+ (id)arrayWithValues:(const void *)values count:(unsigned long)count valueType:(PBArrayValueType)valueType;
++ (id)arrayWithValues:(const void *)values count:(NSUInteger)count valueType:(PBArrayValueType)valueType;
 + (id)arrayWithArray:(NSArray *)array valueType:(PBArrayValueType)valueType;
 - (id)initWithValueType:(PBArrayValueType)valueType;
-- (id)initWithValues:(const void *)values count:(unsigned long)count valueType:(PBArrayValueType)valueType;
+- (id)initWithValues:(const void *)values count:(NSUInteger)count valueType:(PBArrayValueType)valueType;
 - (id)initWithArray:(NSArray *)array valueType:(PBArrayValueType)valueType;
 
 @end
